refactor(webpush): isolate subscriber normalization
